First Colour the Copy Rate button in nice blue colour as we use for text already.
Secondly make sure that we may have more than 1 railway leg and read the data properly. If in the cell we find CY - this means that it is a railway leg. Sometimes we have this as an example: Ж/Д CY Врангель - FOR Селятино. This kind of string must be checked for CY and if there is CY - we write it as railway leg.

After you do these changes please regactor parse.ts , make it in different little functions that work together , so it will be easier later to change or modiy this function.