Let's correct a little part of the boxes on select logic.
We must have a logic that always checks the real available options to choose in the boxes.
For now, we face a little problem with Destination City, please make sure that if we have chosen all the boxes , it will also checks for only avaliably Distiantion City options and show only the existing ones to choose.